corrected error: su2 or u1 cannot be decided at run time. The repo now has 2 programs: ``` su2-main u1-main ``` that have the .cpp sources equal but for the definition of the type "Group" (gauge group)
